pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "happy-sandbox"
version = "0.1.0"
description = "CLI to launch the Claude sandbox Docker environment"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "docopt>=0.6.2",
  "pydantic>=2.0.0",
  "pydantic-settings>=2.0.0",
  "python-socketio>=5.0.0",
  "aiohttp>=3.8.0",
  "httpx>=0.24.0",
  "PyNaCl>=1.5.0",
  "qrcode>=7.0.0",
  "cryptography>=41.0.0",
]

[project.scripts]
happy-sandbox = "happy_sandbox.cli:main"
happy-daemon = "happy_sandbox.daemon:main"
machine-tool = "happy_sandbox.machine_tool:sync_main"

[tool.mypy]
plugins = ["pydantic.mypy"]
check_untyped_defs = true
disallow_untyped_defs = true

[tool.uv]
dev-dependencies = [
  "black>=24.0",
  "mypy>=1.7",
  "pytest>=8.4.2",
]
